{{cookiecutter.project_slug}}/pyproject.toml
[project]
name = "{{ cookiecutter.project_slug }}"
version = "0.1.0"
description = "A Django project"
readme = "README.md"
authors = [
    { name = "{{ cookiecutter.author_name }}", email = "{{cookiecutter.author_email }}" }
]
requires-python = ">={{ cookiecutter.python_version }}"
dependencies = [
    "daphne>=4.2.1",
    "django>=6.0.2",
    "django-cors-headers>=4.9.0",
    "django-dramatiq>=0.11.6",
    "django-environ>=0.12.1",
    "django-jazzmin>=3.0.1",
    "django-ninja>=1.5.3",
    "dramatiq[redis,watch]>=1.17.1",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.11",
    "python-jose[cryptography]>=3.5.0",
    "whitenoise>=6.11.0",
]

[build-system]
requires = ["uv_build>=0.6.0,<0.7.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "toml-cli>=0.8.2",
]
